pyproject.toml
[project]
name = "gapyear_llm"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "starkensin", email = "starkensin@gmail.com" }
]
dependencies = [
    "openai>=1.42.0",
    "python-dotenv>=1.0.1",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/gapyear_llm"]
